+++ devtools/projects/ng-devtools/src/lib/devtools-tabs/directive-explorer/directive-explorer.component.ts
@@ -53,8 +53,8 @@ import {
 import {SplitAreaDirective} from '../../shared/split/splitArea.directive';
 import {SplitComponent} from '../../shared/split/split.component';
 import {Direction} from '../../shared/split/interface';
-import {SignalGraphManager} from './signal-graph/signal-graph-manager';
-import {DevtoolsSignalGraphNode} from './signal-graph';
+import {SignalGraphManager} from './signal-graph-manager/signal-graph-manager';
+import {DevtoolsSignalGraphNode} from '../../shared/signal-graph';
 
 const FOREST_VER_SPLIT_SIZE = 30;
 const SIGNAL_GRAPH_VER_SPLIT_SIZE = 70;
