+++ devtools/projects/ng-devtools/src/lib/devtools-tabs/directive-explorer/property-pane/property-view/property-view.component.ts
@@ -13,7 +13,7 @@ import {ElementPropertyResolver} from '../../property-resolver/element-property-
 import {PropertyViewBodyComponent} from './property-view-body/property-view-body.component';
 import {PropertyViewHeaderComponent} from './property-view-header/property-view-header.component';
 import {FlatNode} from '../../../../shared/object-tree-explorer/object-tree-types';
-import {DevtoolsSignalGraphNode} from '../../signal-graph';
+import {DevtoolsSignalGraphNode} from '../../../../shared/signal-graph';
 
 @Component({
   selector: 'ng-property-view',
