+++ devtools/projects/ng-devtools/src/lib/devtools-tabs/directive-explorer/signal-graph-manager/BUILD.bazel
@@ -1,0 +1,31 @@
+load("//devtools/tools:defaults.bzl", "ng_project", "ts_test_library", "zoneless_web_test_suite")
+
+package(default_visibility = ["//devtools:__subpackages__"])
+
+ng_project(
+    name = "signal-graph-manager",
+    srcs = ["signal-graph-manager.ts"],
+    deps = [
+        "//:node_modules/@angular/core",
+        "//devtools/projects/ng-devtools/src/lib/shared/signal-graph",
+        "//devtools/projects/protocol",
+    ],
+)
+
+ts_test_library(
+    name = "test_lib",
+    srcs = ["signal-graph-manager.spec.ts"],
+    deps = [
+        ":signal-graph-manager",
+        "//:node_modules/@angular/core",
+        "//devtools/projects/ng-devtools/src/lib/shared/signal-graph",
+        "//devtools/projects/protocol",
+    ],
+)
+
+zoneless_web_test_suite(
+    name = "test",
+    deps = [
+        ":test_lib",
+    ],
+)
