+++ devtools/projects/ng-devtools/src/lib/devtools-tabs/directive-explorer/signal-graph-pane/BUILD.bazel
@@ -21,10 +21,11 @@ ng_project(
         "//:node_modules/@angular/material",
         "//devtools/projects/ng-devtools/src/lib/application-operations",
         "//devtools/projects/ng-devtools/src/lib/application-services:frame_manager",
-        "//devtools/projects/ng-devtools/src/lib/devtools-tabs/directive-explorer/signal-graph",
-        "//devtools/projects/ng-devtools/src/lib/devtools-tabs/directive-explorer/signal-graph-pane/signal-details",
-        "//devtools/projects/ng-devtools/src/lib/devtools-tabs/directive-explorer/signal-graph-pane/signals-visualizer",
+        "//devtools/projects/ng-devtools/src/lib/devtools-tabs/directive-explorer/signal-graph-manager",
         "//devtools/projects/ng-devtools/src/lib/shared/button",
+        "//devtools/projects/ng-devtools/src/lib/shared/signal-details",
+        "//devtools/projects/ng-devtools/src/lib/shared/signal-graph",
+        "//devtools/projects/ng-devtools/src/lib/shared/signals-visualizer",
         "//devtools/projects/protocol",
     ],
 )
