+++ devtools/projects/ng-devtools/src/lib/shared/signal-details/signal-details.component.html
@@ -117,6 +117,6 @@
 
 @if (isPreviewable || resourceCluster()) {
   @if (previewableNode(); as previewableNode) {
-    <ng-signal-value-tree [node]="previewableNode" />
+    <ng-signal-value-tree [node]="previewableNode" [element]="element()" />
   }
 }
