+++ devtools/projects/ng-devtools/src/lib/shared/signal-details/signal-details.component.scss
@@ -1,5 +1,5 @@
-@use '../../../../../styles/typography';
-@use '../../../../../styles/theme' as th;
+@use '../../../styles/typography';
+@use '../../../styles/theme' as th;
 
 :host {
   display: block;
