+++ devtools/projects/ng-devtools/src/lib/shared/signal-details/signal-value-tree/signal-data-source.ts
@@ -16,10 +16,10 @@ import {
   MessageBus,
   Properties,
   SignalNodePosition,
-} from '../../../../../../../../protocol';
+} from '../../../../../../protocol';
 import {BehaviorSubject, merge, Observable, Subject} from 'rxjs';
 import {takeUntil} from 'rxjs/operators';
-import {FlatNode, Property} from '../../../../../shared/object-tree-explorer/object-tree-types';
+import {FlatNode, Property} from '../../object-tree-explorer/object-tree-types';
 
 export const arrayifyProps = (
   props: {[prop: string]: Descriptor} | Descriptor[],
