+++ devtools/projects/ng-devtools/src/lib/shared/signal-graph/index.ts
@@ -6,7 +6,6 @@
  * found in the LICENSE file at https://angular.dev/license
  */
 
-export * from './signal-graph-manager';
 export * from './devtools-signal-graph';
 export * from './utils';
 export * from './signal-graph-types';
