+++ devtools/projects/ng-devtools/src/lib/shared/signals-visualizer/signals-visualizer.component.scss
@@ -1,4 +1,4 @@
-@use '../../../../../styles/typography' as tg;
+@use '../../../styles/typography' as tg;
 
 /* Keep in sync with signals-visualizer.ts */
 $node-width: 100px;
