+++ devtools/projects/ng-devtools/src/lib/shared/signals-visualizer/visualizer-types.ts
